See RN4.4.0 (20250714)
- Upgrade aTalk to support Android-16 Baklava (API-36) with the following changes:
  * Android Studio Narwhal | 2025.1.1
  * https://services.gradle.org/distributions/gradle-8.14.2-bin.zip
  * classpath 'com.android.tools.build:gradle:8.11.1'
  * classpath 'org.jetbrains.kotlin:kotlin-gradle-plugin:2.0.21'
- add org.gradle.configuration-cache=false in gradle.properties
- Remove tools:replace="android:allowBackup,android:label" in manifest file.
- add plugins {id 'org.gradle.toolchains.foojay-resolver-convention' version '1.0.0'} in settings.gradle
- Third party libraries upgraded to latest compatible version for API-36:
  * core-ktx: 1.16.0 <= 1.13.1
  * activity-ktx:1.10.1 <= 1.9.3
  * media3-exoplayer:1.7.1 <=> 1.4.1
  * media3-ui:1.7.1 <= 1.4.1
  * threetenabp:1.4.9 <= 1.4.2
  * libphonenumber:9.0.9 <= 8.12.29
  * okhttp:5.1.0 <= 5.0.0-alpha.16
  * commons-text:1.13.1 <= 1.10.0
  * apache-mime4j-core:0.8.12 <= 0.8.9
  * bcpg-jdk15on:1.70, bcpkix-jdk15on:1.70 bctls-jdk15on:1.70 <= 1.65
  * jmdns:3.6.1 <= 3.5.7
  * osmdroid-android:6.1.20 <= 6.1.17
  * lifecycle-runtime-ktx:2.9.1 <= 2.7.0
  * appcompat:1.7.1 < 1.7.0
  * recyclerview:1.4.0 <= 1.3.2
- Build aTalk with support for 16 KB devices; Compile aTalk jni libraries using 16 KB ELF alignment
  * https://developer.android.com/guide/practices/page-sizes#ndk-build
  * use ndkVersion = '28.1.13356709' <= '22.1.7171670'
  * add LOCAL_LDFLAGS += "-Wl,-z,max-page-size=16384" in all Android.mk files
  * add APP_SUPPORT_FLEXIBLE_PAGE_SIZES := true Application.k
- Add support to properly handle EdgeToEdge overlay content in aTalkApp; not compatible with aTalk.
- Received muc message incorrectly classified as outgoing message; introduced in 4.3.0 release.
- Fix auto normal/omemo switching on received omemo message; errors introduced during MenuProvider migration.
- Remove ContentEditText class; Implement ChatController using OnReceiveContentListener interface instead.
- ReImplement FullScreenHelper for aTalk usage and to remove all deprecated functions.
- Replace deprecated ProgressDialog with aTalk customized ProgressDialog implementation.
- Replace deprecated NetworkInfo with networkCapabilities to check for VPN connection in AndroidUsingLinkProperties.
- Migrate deprecated intent.getSerializableExtra() to IntentCompat.getSerializableExtra(), and
  bundle.getSerializable() to BundleCompat.getSerializable().
- Replace TimePicker deprecated method calls.
- ReImplement TouchInterceptor or avoid use of deprecated function calls.
- Reimplement OsmActivity using NetworkCallback instead of BroadcastReceiver.
- Update deprecated function calls in Smack libraries.

*** Observation with no solution implemented:
- AppGUIActivator.getContactListService() (registered in ContactlistActivator) in ChatSessionManager has null reference
  on start debug - SDK JIT problem? not fixed
- JNI static_library_built still using ndk/22.1.7171670; problem with ndk/28.1.13356709
  e.g. llvm-strip: error: 'common/arm/quant-a-10.o': not stripping symbol 'mask_2bit' because it is named in a relocation

+++ aTalk/src/main/java/org/atalk/impl/neomedia/stats/SendTrackStatsImpl.java
@@ -26,8 +26,7 @@ import org.ice4j.util.RateStatistics;
  * @author Boris Grozev
  */
 public class SendTrackStatsImpl extends AbstractTrackStats
-    implements SendTrackStats
-{
+        implements SendTrackStats {
     /**
      * The highest sent sequence number.
      */
@@ -50,29 +49,28 @@ public class SendTrackStatsImpl extends AbstractTrackStats
 
     /**
      * Initializes a new instance.
+     *
      * @param interval the interval in milliseconds over which average bit- and
      * packet-rates will be computed.
      */
-    SendTrackStatsImpl(int interval, long ssrc)
-    {
+    SendTrackStatsImpl(int interval, long ssrc) {
         super(interval, ssrc);
     }
 
     /**
      * Notifies this instance that an RTP packet with a particular sequence
      * number was sent (or is about to be sent).
+     *
      * @param seq the RTP sequence number.
      * @param length the length in bytes.
      */
-    void rtpPacketSent(int seq, int length)
-    {
+    void rtpPacketSent(int seq, int length) {
         long now = System.currentTimeMillis();
 
         // update the bit- and packet-rate
         super.packetProcessed(length, now, true);
 
-        if (highestSeq == -1)
-        {
+        if (highestSeq == -1) {
             highestSeq = seq;
             return;
         }
@@ -84,24 +82,20 @@ public class SendTrackStatsImpl extends AbstractTrackStats
         // between the sender and us, and we need to take this into account
         // when calculating packet loss to the receiver.
         int diff = RTPUtils.getSequenceNumberDelta(seq, highestSeq);
-        if (diff <= 0)
-        {
+        if (diff <= 0) {
             // An old packet, already counted as not send. Un-not-send it ;)
             packetsNotSentRate.update(-1, now);
         }
-        else
-        {
+        else {
             // A newer packet.
             highestSeq = seq;
 
             // diff = 1 is the "normal" case (i.e. we received the very next
             // packet).
-            if (diff > 1)
-            {
+            if (diff > 1) {
                 packetsNotSentRate.update(diff - 1, now);
             }
         }
-
         // update bytes, packets, loss...
     }
 
@@ -113,11 +107,9 @@ public class SendTrackStatsImpl extends AbstractTrackStats
      * (i.e. in the case of jitsi-videobridge the loss rate from the sender to the bridge).
      */
     @Override
-    public double getLossRate()
-    {
+    public double getLossRate() {
         long now = System.currentTimeMillis();
-        if (fractionLostLastUpdate == -1 || now - fractionLostLastUpdate > 8000)
-        {
+        if (fractionLostLastUpdate == -1 || now - fractionLostLastUpdate > 8000) {
             // We haven't received a RR recently, so assume no loss.
             return 0;
         }
@@ -127,8 +119,8 @@ public class SendTrackStatsImpl extends AbstractTrackStats
         long packetsSent = packetRate.getAccumulatedCount(now);
 
         double fractionNotSent
-            = (packetsSent+packetsNotSent > 0)
-            ? (packetsNotSent / (packetsNotSent+packetsSent)) : 0;
+                = (packetsSent + packetsNotSent > 0)
+                ? ((double) packetsNotSent / (packetsNotSent + packetsSent)) : 0;
 
         return Math.max(0, fractionLost - fractionNotSent);
     }
@@ -137,29 +129,28 @@ public class SendTrackStatsImpl extends AbstractTrackStats
      * {@inheritDoc}
      */
     @Override
-    public int getHighestSent()
-    {
+    public int getHighestSent() {
         return highestSeq;
     }
 
     /**
      * Notifies this instance that an RTCP packet with a given length in bytes
      * was sent (or is about to be sent).
+     *
      * @param length
      */
-    void rtcpPacketSent(int length)
-    {
+    void rtcpPacketSent(int length) {
         super.packetProcessed(length, System.currentTimeMillis(), false);
     }
 
     /**
      * Notifies this instance that an RTCP Receiver Report with a given value
      * for the "fraction lost" field was received.
+     *
      * @param fractionLost the value of the "fraction lost" field from an RTCP
      * Receiver Report as an unsigned integer.
      */
-    void rtcpReceiverReportReceived(int fractionLost)
-    {
+    void rtcpReceiverReportReceived(int fractionLost) {
         this.fractionLost = fractionLost / 256d;
         this.fractionLostLastUpdate = System.currentTimeMillis();
     }
